requirements.txt
boto3==1.13.10
PyYAML==5.3.1
jmespath==0.10.0
docutils==0.15.2
